+++ src/hooks.server.ts
@@ -1,5 +1,5 @@
 import type {Handle} from "@sveltejs/kit";
-import {initializeDatabaseConnection, isDatabaseConnectionInitialized} from "$lib/server/database_v3";
+import {initializeDatabaseConnection, isDatabaseConnectionInitialized} from "$lib/server/database";
 
 export const handle: Handle = async ({event, resolve}) => {
     if (!isDatabaseConnectionInitialized())
