+++ src/routes/API/v1/auth/RetrieveCache/+server.ts
@@ -2,7 +2,7 @@ import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
 import {ObjectId} from "mongodb";
 import type {TypeUserCache} from "$lib/stores/UserCache";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const POST = async ({request, cookies}: any): Promise<Response> => {
     consoleLog("RetrieveCache REQUEST Received", LEVEL.OK);
