+++ src/routes/API/v1/farms/GetOneFarmProductsAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const POST = async ({request, cookies}: any): Promise<Response> => {
     consoleLog("GetOneFarmProductsAPI REQUEST Received", LEVEL.OK);
