+++ src/routes/API/v1/farms/GetProductInfoAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const POST = async ({request, cookies}: any): Promise<Response> => {
     consoleLog("GetProductInfoAPI REQUEST Received", LEVEL.OK);
