+++ src/routes/API/v1/forum/GetAllPostAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const GET = async ({cookies}: any) => {
     consoleLog("GetAllPostAPI REQUEST Received", LEVEL.OK);
