+++ src/routes/API/v1/forum/GetMostLikedPostsAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const GET = async ({cookies}: any) => {
     consoleLog("GetMostLikedPostsAPI REQUEST Received", LEVEL.OK);
