+++ src/routes/API/v1/forum/GetOnePostAPI/+server.ts
@@ -1,7 +1,7 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
 import {verifyRequest} from "$lib/server/utility";
 import {ObjectId} from "mongodb";
-import {Database} from "$lib/server/database_v3";
+import {Database} from "$lib/server/database";
 
 export const POST = async ({request, cookies}: any): Promise<Response> => {
     consoleLog("GetOnePostAPI REQUEST Received", LEVEL.OK);
