+++ src/routes/API/v1/news/GetAllNewsAPI/+server.ts
@@ -1,6 +1,6 @@
 import consoleLog, {LEVEL} from "$lib/server/log";
-import {getAllNewsInfo} from "$lib/server/database";
 import {verifyRequest} from "$lib/server/utility";
+import {Database} from "$lib/server/database";
 
 export const GET = async ({cookies}: any) => {
     consoleLog("GetNewsAPI REQUEST Received", LEVEL.OK);
@@ -8,7 +8,7 @@ export const GET = async ({cookies}: any) => {
     if (!verifyRequest(cookies)) {
         return new Response("Unauthorized", {status: 401});
     }
-    const allNewsInfo = await getAllNewsInfo();
+    const allNewsInfo = await Database.getAllNews();
     if (!allNewsInfo) {
         consoleLog("GetNewsAPI RESPONSE: status 404", LEVEL.ERROR);
         return new Response(null, {status: 404});
